requirements.txt
pulsar-client==2.7.3
python-dotenv==0.19.1
ratelimiter==1.2.0.post0
